rgblight: Fix rgblight_blink_layer when multiple layers are active
The change to support repeated blinking (#12237) does not work correctly
when there are multiple blinking layers active at the same time.

It is decrementing the number of times remaining for every active layer
which will cause it to reach 0 prematurely and ignore some of the layers.

Decrement times remaining only once for each timer event.